move conv2d block
